added underline when on discover
